period function amended

+++ test.c
@@ -42,20 +42,20 @@ void pwmsetup(void){
 	pwmWrite(servo_pin, 21);
 }
 
-void rotate(int angle, int anti,int (*peri_deter)(int){
+void rotate(int angle, int anti,double (*peri_deter)(int)){
 	digitalWrite(dir_pin, anti);
 	for(int i = 0; i < angle * MOTOR_STEPS * resolution / 360; i++){
 		digitalWrite(step_pin, HIGH);
-		delayMicroseconds(peri_deter(i));
+		delayMicroseconds((int)(peri_deter(i)));
 		digitalWrite(step_pin, LOW);
-		delayMicroseconds(peri_deter(i));
+		delayMicroseconds((int)(peri_deter(i)));
 	}
 }
 
-int period_determine(int t)
+double period_determine(int t)
 {
 	//uses the variable i (from the for loop in rotate function) to change every time's delay
-	int period_ = 0;
+	double period_ = 0;
 	period_ = e ^ (0.02 * period * t);
 	return period_;
 }
